Split the tileset generation to improve readability
